# src/main/resources/openapi/PrintAPIs.yaml
openapi: 3.0.0
info:
  title: Print APIs
  version: '1.24.0'
  description: Print APIs
#
# --------------------------------------------------------------------------------
#
paths:
...
  '/eros/{eroId}/certificates':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the application.
        schema:
          type: string
        in: path
        required: true
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    get:
      summary: Returns a single certificate summary for a Voter Card Application
      description: Returns a single certificate summary for a Voter Card Application
      parameters:
        - name: applicationId
          description: The identifier of the Voter Card Application.
          schema:
            type: string
          in: query
          required: true
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        '200':
          $ref: '#/components/responses/CertificateSummary'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: get-certificate-summary-by-application-id
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/certificates'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET
  '/eros/{eroId}/certificates/photo':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the application.
        schema:
          type: string
        in: path
        required: true
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    get:
      summary: Returns the Voter Authority Certificate photo for a VAC application
      description: Returns the Voter Authority Certificate photo for a VAC application
      tags:
        - Voter Authority Certificates (VAC)
      parameters:
        - name: applicationId
          description: An identifier of a Voter Authority Certificate application
          schema:
            type: string
          in: query
          required: true
      responses:
        '200':
          $ref: '#/components/responses/PreSignedUrlResource'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: get-voter-authority-certificate-photo-by-application-id
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/certificates/photo'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET
  '/eros/{eroId}/certificates/search':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the Voter Authority Certificates.
        schema:
          type: string
        example: 'city-council'
        in: path
        required: true
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    get:
      summary: Search for Voter Authority Certificates
      description: 'Search for Voter Authority Certificates based on querying parameters'
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        '200':
          $ref: '#/components/responses/CertificateSearchResults'
      operationId: search-voter-authority-certificates
      parameters:
        - name: page
          description: The page number being requested (with the first page being 1). Defaulted to 1 if not supplied
          schema:
            type: integer
            minimum: 1
          example: 1
          in: query
          required: false
        - name: pageSize
          description: The number of results to return in each page. Defaulted to 100 if not supplied
          schema:
            type: integer
            minimum: 1
            maximum: 200
          example: 100
          in: query
          required: false
        - name: searchBy
          description: |
            The data field to apply the `searchValue` to. An empty value will return all VAC documents.
            This value must be provided when `searchValue` is present
          schema:
            $ref: '#/components/schemas/CertificateSearchBy'
          example: applicationReference
          in: query
          required: false
        - name: searchValue
          description: |
            The value to search the field specified by the `searchBy` parameter. 
            The search is a case insensitive match. We do not support fuzzy or wildcard matching.
            This value must be provided when `searchBy` is present
          schema:
            type: string
          in: query
          example: ABCD1X2XX3
          required: false
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/certificates/search'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET
  '/eros/{eroId}/certificates/applications/{applicationId}':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the application.
        schema:
          type: string
        in: path
        required: true
      - name: applicationId
        description: The identifier of the Voter Card Application to retrieve.
        schema:
          type: string
        in: path
        required: true
    options:
      deprecated: true
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    get:
      deprecated: true
      summary: Returns a single certificate summary. Deprecated. Use /eros/{eroId}/certificates?applicationId={applicationId} instead
      description: Returns a single certificate summary. Deprecated. Use /eros/{eroId}/certificates?applicationId={applicationId} instead
      tags:
        - Voter Authority Certificates (VAC)
      responses:
        '200':
          $ref: '#/components/responses/CertificateSummary'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: deprecated-get-certificate-summary-by-application-id
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/certificates/applications/{applicationId}'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.path.applicationId: method.request.path.applicationId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET

  #
  # Voter Authority Certificate Statistics - anonymised statistics for a VAC
  # --------------------------------------------------------------------------------
  '/certificates/statistics':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Voter Authority Certificates (VAC)
        - Statistics
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    get:
      summary: Returns anonymised statistics for a Voter Authority Certificate application
      description: Returns anonymised statistics for a Voter Authority Certificate application
      tags:
        - Voter Authority Certificates (VAC)
        - Statistics
      parameters:
        - name: applicationId
          description: An identifier of an Voter Authority Certificate application
          schema:
            type: string
          in: query
          required: true
      responses:
        '200':
          $ref: '#/components/responses/Statistics'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: get-certificate-statistics-by-application-id
      security:
        - awsIamAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/certificates/statistics'
        requestParameters:
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET


  #
  # Temporary Certificates for a Voter Card Application
  # --------------------------------------------------------------------------------
...
components:
...
  schemas:
    ErrorResponse:
      title: ErrorResponse
      description: |
        A generic response body object describing errors in a web request, and can be used to communicate several
        different types of error condition such as (but not limited to) `400 BAD REQUEST`, `409 CONFLICT` etc.
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2022-09-28T18:01:42.105Z'
        status:
          type: integer
          example: 400
        error:
          type: string
          example: 'Bad Request'
        message:
          type: string
          example: 'Validation failed for object=applicationRequest. Error count: 14'
        validationErrors:
          description: Validation errors are only present if the error being described is a `400 BAD REQUEST`.
          type: array
          items:
            type: string
          example: 'Error on field applicant.nino: rejected value [aaaaaaaaaaa], must match ^.{1,10}$'
      required:
        - timestamp
        - status
        - error
        - message
    CertificateSummaryResponse:
      title: CertificateSummaryResponse
      type: object
      description: A response containing a single Certificate summary
      properties:
        vacNumber:
          type: string
          pattern: '^[0-9AC-HJ-NP-RT-Z]{20}$'
          description: The unique identifier of the Certificate. A 20 character alphanumeric string.
          example: DEV1W0XDH1368LWE1J40
        applicationReference:
          type: string
          example: V3JSZC4CRH
          description: The application reference as known by the citizen. Not guaranteed to be unique.
        sourceReference:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          description: |
            A 24 character hex string. 
            The unique identifier of a Voter Authority Certificate application.
          example: 1f0f76a9a66f438b9bb33070
        firstName:
          type: string
          maxLength: 255
          description: First name of the Elector
          example: John
        middleNames:
          type: string
          maxLength: 255
          description: Middle names of the Elector
          example: Malcolm
        surname:
          type: string
          maxLength: 255
          description: Surname of the Elector
          example: Smith
        printRequestSummaries:
          type: array
          description: List of print request summaries for the certificate.
          items:
            $ref: '#/components/schemas/PrintRequestSummary'
        photoUrl:
          type: string
          description: |
            The URL of the Elector's Certificate photo. 
            This is the photo that has been prepared and approved for the Certificate by the ERO, rather than the photo that the applicant uploaded as part of their original application.
          example: 'http://localhost:8080/eros/city-council/certificates/photo?applicationId=507f1f77bcf86cd799439011'
      required:
        - vacNumber
        - applicationReference
        - printRequestSummaries
        - photoUrl
    CertificateSearchSummaryResponse:
      title: CertificateSearchSummaryResponse
      description: Response containing the meta data and results of searching for Voter Authority Certificate summaries
      type: object
      properties:
        page:
          type: integer
          description: The page number
          example: 1
        pageSize:
          type: integer
          description: The number of results per page
          example: 1
        totalPages:
          type: integer
          description: The total number of pages
          example: 1
        totalResults:
          type: integer
          description: The total number of results
          example: 1
        results:
          type: array
          description: Array of voter authority certificate search summaries that matches the search criteria
          items:
            $ref: '#/components/schemas/CertificateSummaryResponse'
      required:
        - page
        - pageSize
        - totalPages
        - totalResults
        - results
    PrintRequestSummary:
      title: PrintRequestSummary
      description: An object containing print request summary
      type: object
      allOf:
        - $ref: '#/components/schemas/UpdatedBy'
      properties:
        status:
          $ref: '#/components/schemas/PrintRequestStatus'
        message:
          type: string
          description: Message received for the print request status update
        deliveryAddressType:
          $ref: '#/components/schemas/DeliveryAddressType'
      required:
        - status
    PrintRequestStatus:
      title: PrintRequestStatus
      type: string
      enum:
        - print-processing
        - print-failed
        - dispatched
        - not-delivered
    CertificateStatisticsStatus:
      title: CertificateStatisticsStatus
      type: string
      enum:
        - PENDING_ASSIGNMENT_TO_BATCH
        - ASSIGNED_TO_BATCH
        - SENT_TO_PRINT_PROVIDER
        - RECEIVED_BY_PRINT_PROVIDER
        - VALIDATED_BY_PRINT_PROVIDER
        - IN_PRODUCTION
        - DISPATCHED
        - NOT_DELIVERED
        - PRINT_PROVIDER_VALIDATION_FAILED
        - PRINT_PROVIDER_PRODUCTION_FAILED
        - PRINT_PROVIDER_DISPATCH_FAILED
...
    CertificateSearchBy:
      title: CertificateSearchBy
      description: An enumeration of search by fields for searching VACs
      type: string
      enum:
        - surname
        - applicationReference
...
    UpdatedBy:
      title: UpdatedBy
      description: Object containing the updated by details of a resource.
      type: object
      properties:
        userId:
          description: "UserId (email address) of who last updated the resource."
          format: email
          type: string
          example: fred@some-ero.gov.uk
        dateTime:
          description: "The timestamp in ISO-8601 UTC when an update on the resource occurred."
          type: string
          format: date-time
          example: '2022-08-17T16:38:06Z'
      required:
        - userId
        - dateTime
...
    DeliveryAddressType:
      title: DeliveryAddressType
      description: The delivery address type for the Anonymous Elector Document and Voter Authority Certificate
      type: string
      enum:
        - registered
        - ero-collection
        - alternative
...
    StatisticsResponse:
      title: StatisticsResponse
      type: object
      description: A response containing anonymised statistics for a VAC or AED application.
      properties:
        certificateStatus:
          $ref: '#/components/schemas/CertificateStatisticsStatus'
        temporaryCertificateIssued:
          type: boolean
        certificateReprinted:
          type: boolean
      required:
        - certificateStatus
        - certificateReprinted
        - temporaryCertificateIssued
...
    PreSignedUrlResourceResponse:
      title: PreSignedUrlResourceResponse
      description: An object detailing the pre-signed URL to use in order to retrieve the resource
      properties:
        preSignedUrl:
          type: string
          format: uri
          description: AWS pre-signed URL giving time-limited access to the resource.
          example: 'https://example.s3.eu-west-2.amazonaws.com/image.png?X-Amz-Expires=1800'
      required:
        - preSignedUrl

  #
  # Response Body Definitions
  # --------------------------------------------------------------------------------
  responses:
...
    404ErrorResponse:
      description: |
        Error response for a HTTP 404 describing the problem with the request.
        Typical examples are resources such as the application not found.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2022-09-28T18:01:42.105Z'
            status: 404
            error: 'Not Found'
            message: 'Certificate photo for voter card application with eroId=[some-ero] and applicationType=[VOTER_CARD] and applicationId=[507f1f77bcf86cd799439011] not found'
...
    CertificateSummary:
      description: Response containing a single Certificate summary
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CertificateSummaryResponse'
...
    CertificateSearchResults:
      description: Response containing the search result of Voter Authority Certificates
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CertificateSearchSummaryResponse'
    Statistics:
      description: Response containing statistics for an application
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StatisticsResponse'
    PreSignedUrlResource:
      description: Response containing an object detailing the pre-signed URL for the resource
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PreSignedUrlResourceResponse'

  #
  # Request Body Definitions
  # --------------------------------------------------------------------------------
...
  securitySchemes:
    eroUserCognitoUserPoolAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        providerARNs:
          - '${cognito_ero_user_pool_arn}'
        type: cognito_user_pools
    awsIamAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: awsSigv4